Feat: #12 공용시설 예약

+++ backend/src/main/java/com/ohammer/apartner/domain/facility/repository/FacilityRepository.java
@@ -3,6 +3,6 @@ package com.ohammer.apartner.domain.facility.repository;
 import com.ohammer.apartner.domain.facility.entity.Facility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacilityRepository extends JpaRepository<Facility, Integer> {
+public interface FacilityRepository extends JpaRepository<Facility, Long> {
 
 }
